input large price numbers without commas
The input masks (which prevented the input of non-digit glyphs) started to create a problem. When you input a large number as the price and then go back to edit it, the price would show up with commas with the standard NumberStringConverter class. I replace this with a simpler doubleStringConverter which does the straightforward thing: printf("%.2f") and Double.valueOf(x).